requirements.txt
Django==3.2.14
djangorestframework==3.13.1
djoser==2.1.0
django-cors-headers==3.10.1